+++ policy-engine/src/main/java/com/walmartlabs/concord/policyengine/WorkspacePolicy.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.policyengine;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -31,6 +31,7 @@ import java.nio.file.attribute.BasicFileAttributes;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Set;
 
 public class WorkspacePolicy {
 
@@ -47,29 +48,45 @@ public class WorkspacePolicy {
 
         List<CheckResult.Item<WorkspaceRule, Path>> deny = new ArrayList<>();
 
-        if (!Files.exists(p) || !Files.isDirectory(p)) {
-            // TODO pass additional error information
-            deny.add(new CheckResult.Item<>(rule, p));
-        }
-
-        if (rule.getMaxSizeInBytes() != null) {
+        if (!Files.exists(p)) {
+            deny.add(new CheckResult.Item<>(rule, p, "File not found: " + p));
+        } else if (!Files.isDirectory(p)) {
+            deny.add(new CheckResult.Item<>(rule, p, "Not a directory: " + p));
+        } else if (rule.getMaxSizeInBytes() != null) {
             MutableLong size = new MutableLong();
             Files.walkFileTree(p, new SimpleFileVisitor<Path>() {
+
                 @Override
                 public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-                    if (Files.isRegularFile(file)) {
-                        size.add(Files.size(file));
+                    if (isIgnored(file, rule.getIgnoredFiles())) {
+                        return FileVisitResult.CONTINUE;
                     }
+
+                    size.add(Files.size(file));
 
                     return FileVisitResult.CONTINUE;
                 }
             });
 
             if (size.toLong() > rule.getMaxSizeInBytes()) {
-                deny.add(new CheckResult.Item<>(rule, p));
+                deny.add(new CheckResult.Item<>(rule, p, "Workspace too big: " + size + " byte(s)"));
             }
         }
 
         return new CheckResult<>(Collections.emptyList(), deny);
+    }
+
+    private static boolean isIgnored(Path p, Set<String> patterns) {
+        if (patterns == null) {
+            return false;
+        }
+
+        for (String s : patterns) {
+            if (p.toString().matches(s)) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }